Refactor: send components to components folder
