+++ src/app/components/characters/characters.component.scss
