+++ src/app/components/individual/individual.component.html
