+++ src/app/components/individual/individual.component.scss
